removed unnecessary casts (int)

+++ ExampleAIModule/Source/TaskEngine.cpp
@@ -2,10 +2,10 @@
 
 void assessTask(array<int, 12>& myTask, int frameCount, int minerals, int gas, Unit CommandCenter, list<int>& Miners, list<int>& Builders)
 {
-    int taskStatus = myTask[(int)tsk::Status];
-    int callBackTime = myTask[(int)tsk::TimeStamp] + myTask[(int)tsk::Delay];
+    int taskStatus = myTask[tsk::Status];
+    int callBackTime = myTask[tsk::TimeStamp] + myTask[tsk::Delay];
 
-    if (taskStatus == (int)taskStatus::Created || taskStatus == (int)taskStatus::waitingMin)
+    if (taskStatus == taskStatus::Created || taskStatus == taskStatus::waitingMin)
     {
         if (frameCount > callBackTime)
         {
@@ -14,29 +14,29 @@ void assessTask(array<int, 12>& myTask, int frameCount, int minerals, int gas, U
                 //assign SCV, instead of SCV just get the ID directly, remove BWAPI 
                 Unit SCV = UnitFun::getWorker(CommandCenter, Miners, Builders);
                 //pass the worker id to the task
-                myTask[(int)tsk::UID] = SCV->getID();
+                myTask[tsk::UID] = SCV->getID();
 
-                if (myTask[(int)tsk::X] == 0 && myTask[(int)tsk::Y] == 0)
+                if (myTask[tsk::X] == 0 && myTask[tsk::Y] == 0)
                 {
-                    const TilePosition targetBuildLocation = BuildManager::returnBuildPosition(myTask[(int)tsk::Action], SCV, 20);
-                    myTask[(int)tsk::X] = targetBuildLocation.x;
-                    myTask[(int)tsk::Y] = targetBuildLocation.y;
+                    const TilePosition targetBuildLocation = BuildManager::returnBuildPosition(myTask[tsk::Action], SCV, 20);
+                    myTask[tsk::X] = targetBuildLocation.x;
+                    myTask[tsk::Y] = targetBuildLocation.y;
                 }
                 TaskFun::startTask(myTask, SCV);//set to PendingStart, onUnitCreate set to Start, onUnitComplete set to Complete
             }
             else
             {
                 //do we have resources? assign, no? set callback time
-                TaskFun::callBack(myTask, 200, (int)taskStatus::waitingMin);
+                TaskFun::callBack(myTask, 200, taskStatus::waitingMin);
             }
         }
     }
-    else if (taskStatus == (int)taskStatus::PendingStart)
+    else if (taskStatus == taskStatus::PendingStart)
     {
         if (frameCount > callBackTime)
         {
-            myTask[(int)tsk::Status] = (int)taskStatus::Cancelled;
-            Broodwar->sendText("TaskMngr: Failed to start task id: %d : ", myTask[(int)tsk::ID]); //remove the log to broodwar, send a error code
+            myTask[tsk::Status] = taskStatus::Cancelled;
+            Broodwar->sendText("TaskMngr: Failed to start task id: %d : ", myTask[tsk::ID]); //remove the log to broodwar, send a error code
         }
     }
     else
@@ -56,10 +56,10 @@ void TaskEngine::taskManager(list<array<int, 12>>& myTaskQueue, int frameCount, 
     for (auto& task : myTaskQueue)
     {
         //assessTask(task,frameCount,minerals,gas,CommandCenter,Miners,Builders);
-        int taskStatus = task[(int)tsk::Status];
-        int callBackTime = task[(int)tsk::TimeStamp] + task[(int)tsk::Delay];
+        int taskStatus = task[tsk::Status];
+        int callBackTime = task[tsk::TimeStamp] + task[tsk::Delay];
 
-        if (taskStatus == (int)taskStatus::Created || taskStatus == (int)taskStatus::waitingMin)
+        if (taskStatus == taskStatus::Created || taskStatus == taskStatus::waitingMin)
         {
             if (frameCount > callBackTime)
             {
@@ -68,29 +68,29 @@ void TaskEngine::taskManager(list<array<int, 12>>& myTaskQueue, int frameCount, 
                     //assign SCV, instead of SCV just get the ID directly, remove BWAPI 
                     Unit SCV = UnitFun::getWorker(CommandCenter, Miners, Builders);
                     //pass the worker id to the task
-                    task[(int)tsk::UID] = SCV->getID();
+                    task[tsk::UID] = SCV->getID();
 
-                    if (task[(int)tsk::X] == 0 && task[(int)tsk::Y] == 0)
+                    if (task[tsk::X] == 0 && task[tsk::Y] == 0)
                     {
-                        const TilePosition targetBuildLocation = BuildManager::returnBuildPosition(task[(int)tsk::Action], SCV, 20);
-                        task[(int)tsk::X] = targetBuildLocation.x;
-                        task[(int)tsk::Y] = targetBuildLocation.y;
+                        const TilePosition targetBuildLocation = BuildManager::returnBuildPosition(task[tsk::Action], SCV, 20);
+                        task[tsk::X] = targetBuildLocation.x;
+                        task[tsk::Y] = targetBuildLocation.y;
                     }
                     TaskFun::startTask(task, SCV);//set to PendingStart, onUnitCreate set to Start, onUnitComplete set to Complete
                 }
                 else
                 {
                     //do we have resources? assign, no? set callback time
-                    TaskFun::callBack(task, 200, (int)taskStatus::waitingMin);
+                    TaskFun::callBack(task, 200, taskStatus::waitingMin);
                 }
             }
         }
-        else if (taskStatus == (int)taskStatus::PendingStart)
+        else if (taskStatus == taskStatus::PendingStart)
         {
             if (frameCount > callBackTime)
             {
-                task[(int)tsk::Status] = (int)taskStatus::Cancelled;
-                Broodwar->sendText("TaskMngr: Failed to start task id: %d : ", task[(int)tsk::ID]); //for some error 
+                task[tsk::Status] = taskStatus::Cancelled;
+                Broodwar->sendText("TaskMngr: Failed to start task id: %d : ", task[tsk::ID]); //for some error 
             }
         }
         else
